pyproject.toml
[project]
name = "heimdall"
version = "0.1.0"
description = "Autonomous agentic machine administration"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0",
]

[project.scripts]
heimdall = "heimdall.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/heimdall"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
